Elegoo_Car_Project/requirements.txt
numpy
opencv-python
PyYAML
requests
pupil-apriltags
imageio
imageio-ffmpeg
matplotlib
keyboard
